requirements.txt
flask
pymysql
python-dotenv 
flask-restx